refactored prject structure of the infrastructure persistence layer
