+++ Infrastructure/Persistence/LibSystem.Persistence/Repositories/BorrowingRepository.cs
@@ -1,6 +1,7 @@
 ﻿using LibSystem.Application.Repositories;
 using LibSystem.Domain.Entities.Borrowing;
 using LibSystem.Domain.ValueObjects;
+using LibSystem.Persistence.Context;
 using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
