+++ Infrastructure/Persistence/LibSystem.Persistence/Repositories/GenericRepository.cs
@@ -1,4 +1,5 @@
 ﻿using LibSystem.Domain.Common;
+using LibSystem.Persistence.Context;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Storage;
 using System;
